Merge multiple CSP headers
Merging is not technically correct. A CSP policy is a set of directives.
Duplicate directives are ignored. In case of multiple policies,
"adding additional policies to the list of policies to enforce can only
further restrict the capabilities of the protected resource"
(https://w3c.github.io/webappsec-csp/#multiple-policies).

Still, treating it as one long policy (as this commit does) is probably
less broken than what we did before. "Temporary" workaround until we get
something better in place...